use selectinput
